+++ app/(auth)/new-password/page.tsx
@@ -1,7 +1,12 @@
 import { NewPasswordForm } from '@/features/auth/components/new-password-form';
+import { Suspense } from 'react';
 
 const NewPasswordPage = () => {
-  return <NewPasswordForm />;
+  return (
+    <Suspense fallback={<div>Loading...</div>}>
+      <NewPasswordForm />
+    </Suspense>
+  );
 };
 
 export default NewPasswordPage;
